Don't pass INITD to make install-redhat in dkms.spec
Reported by Charles Rose.  When the spec was rpmbuilt on Red Hat, this
causes it to put the file in /etc/rc.d/init.d.  But that directory
doesn't exist on SLES9 and SLES10.  Instead, we always want it to go
in /etc/init.d (which on Red Hat is a symlink), so that works for all
Red Hat and SuSE products.