uninstall from /extra before DEST_MODULE_LOCATION
https://bugzilla.redhat.com/show_bug.cgi?id=264981

It only fails when uninstalling something for which there's already
one elsewhere in the tree - e.g. in both kernel/fs/fuse and extras/.
Turns out, this function here was returning the lookup in the wrong
order, so it would delete the one in kernel/fs/fuse rather than
deleting the one in extras/ like it should.  It needs to delete from
extras/ first.