Add patch from Andy Whitcroft:
In addition to the fixes for the kernel packages to clean themselves
up correctly, DKMS is also making a mess and not cleaning
up when all its modules are removed.  Specificially it makes a
/lib/modules/<version>/updates/dkms directory to hold the installed
files.
When they are removed this directory is not removed again.  This leads
dpkg to leave this directory as it is 'not empty'.

Below is a debdiff to the DKMS kernel prerm triggers to attempt to clear
these out when dkms is cleaning up for a kernel which is being removed.

-apw