Add in another patch from Robbie Basak:
Previously, -m and -v were required in order to identify the module
being worked on.

Now, if neither -m nor -v are supplied, it is assumed that the current
directory is a working copy of a DKMS source tree. A temporary source
tree is created by taking a copy of the current directory, a temporary
DKMS tree is created, the source tree added, and a DKMS package
generated in the parent directory.

For example, from a working copy of a DKMS source tree:
    dkms mkdeb --source-only

Currently only the mkdeb and mkdsc commands are supported for use in
this fashion, but a framework exists to add support for other commands
in the future.