Add a patch from Thomas Chenault:
Loading modules from a SLES driver disk may fail in cases where the
modules on the driver disk are dependent on modules provided by the OS.
The attached patch causes dkms to automatically create driver disk
images that induce loading of the required modules.

In cases where modules that are not provided on the driver disk are
required, the patched dkms will create a driver disk with a layout
similar to the following.

/media/floppy/
|-- 01
|   `-- linux
|       `-- suse
|           `-- x86_64-sles11
|               `-- modules
|                   `-- module.order
`-- 02
    `-- linux
        `-- suse
            `-- x86_64-sles11
                |-- install
                |   |-- update.post
                |   `-- update.tar.gz
                `-- modules
                    `-- igb.ko

The first update on the disk contains no modules, only a module.order
file which lists dependencies. Testing has shown this to be an effective
way to cause the needed module(s) to load. The second and subsequent
updates on the disk carry the actual payload.