Ditch is_word_in_list and bashify etc_sysconfig_kernel_modify.
is_word_in_list is only used in one place, and is is easier to just inline it.

Return/continue early instead of enclosing large blocks of code in if.

Make the whole etc_sysconfig_kerne_modify function run in a subshell.
Source /etc/sysconfig/kernel instead of doing the eval $(grep) thing.