Bashify the rest of dkms
Also make 2 helper functions -- check_root and read_arg.

check_root will die with an error message if we are not running as root.

read_arg does The Right Thing when reading and interning args
that can either be:
	 --foo bar
	 --foo=bar

This ditches lots of boilerplate -- these along with formatting changes
cut over 200 lines out of the code.