Add new syntax for specfying module/modver and kernelver/arch
As you may have guessed, you can now pass these as a single string
with a slash between the components instead of as seperate args.
The internal logic is unchanged for now.

Also, replace gratuitous use of find with globbing in do_status_weak.
This should help speed things up a bit.