CLI interface usage improvments.
Allow the user to specify the module and module version without having
to use the -m and -v switches.  Instead, leverage parse_moduleversion
to set the module and module_version if we are passed a non-switch arg
that is not an action.

This makes it possible to use command lines like

dkms status tg3
dkms add netxtreme2-5.2.28

To accomidate module names that have a - in them. we can also pass
module/module_version, like so:

dkms add netxtreme2/5.2.28

The -m switch also uses parse_moduleversion, so you can use it the same way.