Merge skip-non-matching-kernel patch from Debian
dkms.conf allows to set a BUILD_EXCLUSIVE_KERNEL / BUILD_EXCLUSIVE_ARCH
parameter to specify a regular expression which defines the subset of
kernels / architectures which DKMS is allowed to build the module for. If
the kernel being built for does not match against this regular
expressions, the dkms build will error out with exit code 9.

This patch changes dkms_common.postinst to handle this dkms build
error gracefully by skipping build for the non matching kernels.