Fix add_source_tree if the destination directory exists
If "$source_tree/$module-$module_version" already exists when
add_source_tree is called, the cp command needs to be aware of this so
it only copies the contents of "$from" and not the containing directory
as well.

This may occur if the tarball passed to ldtarball happens to be stored
in /usr/src/$module-$module_version/

By always creating the destination directory, the cp command can assume
it is always present, and succeed whether or not the destination
directory was already present.

Signed-off-by: Jim Ramsay <jim_ramsay@dell.com>