Add a patch from Charlie Brooks to fix some missing undefined local variables.
Like most subroutines in dkms, module_status_weak() uses the variables
m, v, k, and a, but unlike the others it does not declare them local.
This pollutes the variables in the caller based on the content of the
DKMS tree, with highly unpredictable results in any specific
installation.

Tested on RHEL6 kernels 2.6.32-220.17.1.el6.i686 and
2.6.32-279.el6.i686, 2012-06-27 CTB

Signed-off-by: Charlie Brooks <medievalist@gmail.com>