Add a patch from Charlie Brooks:
The case statement at the end of /usr/sbin/dkms has a case for a null
$action_to_run, but $action_to_run is set by a do loop on the content
of $action, so it's impossible for $action_to_run to be set null
unless you're doing foolish things with metacharacters and spaces on
the command line, and thus there's an untrapped common error condition
that exits with a SUCCESS value and no output.

The following patch solves the problem in the coarsest sense, since it
supplies a FAIL exit status.  The parse_moduleversion function doesn't
catch non-existent modules, though, so the error text isn't optimally
useful.

Signed-off-by: Charlie Brooks <medievalist at gmail.com>