Clear variable res on get_module_verinfo()
When building a DKMS module that don't supplies MODULE_VERSION strings
that tries to replace the in-kernel module that supplies this
MODULE_VERSION string DKMS fails to recognize that both modules are
different because it is not cleaning the variable res.

This causes that DKMS don't installs the new module that is clearly
different than the in-kernel one (the first has no MODULE_VERSION and
the second one has).

More info on this Debian bug: http://bugs.debian.org/684326

Signed-off-by: Carlos Alberto Lopez Perez <clopez@igalia.com>