Do not unnecessarily sleep in multiples of three seconds
We are not in the Pentium 4 era, a delay of three seconds is unacceptable these
days. First run a command backgrounded, then print progress dots as long as
the command is running (/proc is assumed to be mounted). Then wait for
completion of the command and save its exit code. When done, kill the progress
dots to avoid printing a final dot. The final wait for the progress dots exists
to avoid messages like:

    dkms: line 74: 123 Terminated while [ -d /proc/$pid ]; do

While using such wait constructs, there is no need for a temporary file to store
the exit code, so remove that too.