Fix removal of module build directory
The path expansion did not work because the * was inside the quoted
string. The fix removes the path expansion alltogether as the whole
build directory is removed before every build anyway. So there is no
need to keep it around.

For larger modules or modules including binary code (like the non-free
graphics drivers) this saves quite some disk space which won't be used
on a subsequent build anyway.