Fix check for obsolete modules
The check did not work as the array modules_conf_obsoletes array may
contain elements even if there are no obsoleted modules. This is due to
how the array is initialized. It's always as large as the largest of the
configuration arrays defined in dkms.conf.

This fix does a string comparison instead. If no osbsolete module is
defined, then the expansion of the whole array to a string is always
empty. For some reason [[ -n ${{modules_conf_obsoletes[@]} ]] does not
work here.