Allow building for all installed kernels in the common postinst
This adds a configuration option to framework.conf to allow building for
all installed kernels instead of the currrent and latest one. This is
especially usefull for live systems which have several kernels installed
for different hardware configurations (ie. PAE and non-PAE kernels). In
this situation it's desirable to build the DKMS modules for all
installed kernels as you are quite likely to boot into other kernels on
different machines. By default this option is set to off.

This commit is loosely based on a patch found in the Debian bug tracking
system at https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=710535

It also reworks some of the kernel selection code in general to contain
less special cases and to not have an own loop for checking the current
kernel. This rework should not contain any functional changes beside
this new configuration setting.