Bash 4.3 syntax changes.
According to man pages, ${arr[@]} expands the array into separate words (analogous to the expansion of $@). Quoting the man page: 'That is, "$@" is equivalent to "$1" "$2" ...' So what it basically becomes in your example is <value-of-arr0><separator><value-of-arr1>.

Substituting the relevant piece of code with [[ \"${arr[@]}\\" ]] demonstrates this well.

There are minor distinctions in expanding using * and @.
The man page states:
"If the word is double-quoted, ${name[*]} expands to a single word with the value of each array member separated by the first character of the IFS special  variable".