Preserve timestamps when copying source into build directory
Trying to create DKMS-aware RPM, we trigger an issue that we identified to be
caused by the way DKMS copies source into its build directory which leads to
files timestamps not being preserved and thus some files dependencies can be
triggered and files rebuilt by mistake. This may finally lead to unexpected
errors when required tools for such files rebuild are missing on the target
system.

Reported:

- https://bugs.launchpad.net/dkms/+bug/952817
- https://bugzilla.redhat.com/show_bug.cgi?id=912300