postinst: Fix newest kernel detection
We were passing a list of kernels to get_newest_kernel(), when it
expected a single kernel as its first argument - the currently running
kernel. This in turn meant that the sometimes (if the first thing in the
list was itself the newest kernel), the wrong kernel was passed into
_get_newest_kernel_debian(). As a result this falsely reported that
there was no 'newest' kernel available and then didn't build modules for
it.

Drop the parameter to get_newest_kernel(), and instead rely on
CURRENT_KERNEL being set to $(uname -r).

https://bugs.launchpad.net/ubuntu/+source/dkms/+bug/1681566