Fix syntactic error when the system shell is zsh
When the system '/bin/sh' points to the zsh shell,
then the script is run with zsh in sh compatibility mode; in that case
functions must end with either a semicolon or a newline.
(https://github.com/dell/dkms/issues/62)
* dkms_autoinstaller (log_end_msg):
Terminate the function with a semicolon.
Add missing white space in comment.