Apply patch from Debian: add mkbmdeb support for lean binary package
https://salsa.debian.org/debian/dkms/blob/master/debian/patches/0004-mkbmdeb-support-for-lean-binary-package-with-only-th.patch

Patch comment:

From: Thijs Kinkhorst <thijs@debian.org>
Date: Wed, 6 Jul 2016 05:34:43 +0800
Subject: mkbmdeb: support for lean binary package with only the built
modules

Creates a Debian binary package containing just the binary modules in
the
/lib/modules installation path. This package does not depend on dkms
and
does not require a toolchain to be installed on the target host. Useful
if you want to have a package to install on hosts identical to the
build
system without installing the full toolchain on them.
Bug-Debian: http://bugs.debian.org/554843