Explicitly check for --all with add/build/install actions
Using --all with add, simply cannot work: the presence of the module is
used to determine the list of "all" kernels/arches.

Technically it is possible to have --all working with build/install,
when the module is already added.

Doing that is a bit more involving than the time I have at the moment,
so for now explicitly error out. The option is not supported, as per the
manual.

Note: the option --all in not allowed with a few other actions. I've
left those out for now since they're not ones I could realistically test

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>