dkms_common.postinst: simplify _is_kernel_name_correct
This patch removes the _is_kernel_name_correct function's dependency on
on the existence of the /boot/config-* files. Instead, of keying on the
existence of the /boot/config-* files, this patch changes the function
to key on the existence of the /lib/modules/* directories. This change
is necessary for modern distributions such as Fedora 33 that have
switched to the newer Boot Loader Specification (BLS). Systemd-boot, in
particular, does not create the /boot/config-* files when new kernels
are added with its kernel-install script.

Signed-off-by: Gregory Bartholomew <gregory.lee.bartholomew@gmail.com>