make: consistently handle input files - dkms.in dkms.8.in
Currently the installed files are being manipulated, which is fairly
uncommon and strange.

The correct/consistent way of handling it, is to rename/suffix our
dkms{,.8} with .in, and do the manipulation locally before everything
else.

As an added bonus, we now fix the release string for dkms.8 in the
tarball.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>