Support CC=clang and LD=ld.lld
When building the Linux kernel, one may use `make LLVM=1` to set
multiple flags, akin to `make CC=clang LD=ld.lld NM=llvm-nm ...` (see
the link below for kernel docs explaining this further in detail).

When building kernel modules, to ensure we're using the same toolchain
as the underlying core kernel image, Kbuild will error if it's reinvoked
with a toolchain that differs. This causes DKMS to fail, since it's not
re-specifying the same compiler (or linker, etc).

Check the .comment section of the vmlinux file in the Linux kernel
source dir, and set CC= and LD= flags for make based on that.

If vmlinux does not exist, grep .config for CONFIG_CC_IS_CLANG=y and
CONFIG_LD_IS_LLD=y.

Fixes: #124
Fixes: https://github.com/ClangBuiltLinux/linux/issues/1104
Link: https://docs.kernel.org/kbuild/llvm.html
Suggested-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Nick Desaulniers <nick.desaulniers@gmail.com>