{ld,mk}tarball: perform expansion when setting cleanup trap
ldtarball stored the name of the temporary directory where the
archive was to be extracted to in a local variable, and set a
trap to clean up the directory upon exit. The trap was set with
single quotes, meaning that the value of $temp_dir_name was
expanded when the trap was run, at which point the variable was
out of scope.

Change the trap to use double quotes, which expands the variable
when the trap is set, allowing the cleanup of the ldtarball
staging directory to work correctly. mktarball also used single
quotes, but its cleanup trap worked correctly because mktarball's
temp_dir_name variable was not locally scoped. For consistency,
change mktarball's temp_dir_name to a local variable and set the
trap with double quotes.

For dell/dkms#190: ldtarball doesn't clean up after itself

Signed-off-by: Daniel Dadap <ddadap@nvidia.com>