sign_build: prefer scripts/sign-file to kmodsign
The (really useful) sign_build function was introduced a while ago in
b25cb7e4668e45104e77ed65d43360743668d51e, and it relies on kmodsign to
sign kernel modules so that they can be used in a Secure Boot environment.

Unfortunately, kmodsign only exists in Ubuntu, so this function was
quite useless on other distros.

In this commit I made it possible to instead use scripts/sign-file:
https://www.kernel.org/doc/html/latest/admin-guide/module-signing.html#manually-signing-modules
They do exactly the same thing, as kmodsign is based on the same code
and adds just a couple of extra options, but sign-file is included in
the upstream kernel and is more broadly available.

To avoid potentially breaking existing setups, kmodsign is kept as a
fallback.

sign_build() still makes use of update-secureboot-policy, a tool that
exists only on Debian and Ubuntu, so this patch currently improves the
situation only for Debian.

I've also made the function a bit more verbose, addressing Gianfranco's
concerns when submitting the patch:
https://github.com/dell/dkms/pull/104#discussion_r339418815