Put version comparison before info comparison
The priority of version comparison should be higher than
the priority of module content comparison. Sometimes when
we need to fix some bugfixes, there may be no difference
when comparing module content, resulting in dkms not being
able to be installed and still keeping the drivers in the
original system. If we compare the versions first, and if the
version is higher than the driver version in the current system,
we replace it and then can solve this problem.

Signed-off-by: Meng Tang <tangmeng@uniontech.com>