dkms.in: skip sign_file call when unset
The module-signing script is located in prepare_signing(). It is
possible for installation to proceed, even if no signing file is located
on the system. In that case, the script promises that signing will not
occur.

However, in actual_build(), there is no check for the case where
$sign_file is unset. So `eval "" sha512 ...` is executed instead - which
is a nonsense command.

If $sign_file is unset, skip signing.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>