run-test.sh: Split logic to genericize expected output into helper.
The logic to genericize the expected output via sed was
previously only used for run_with_expected_output and
not by run_with_expected_error. Splitting that into a
helper method allows it to be used in both places.

This change also uses a variable for the log file names
so that only one place needs updated in each helper.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>