run-test.sh: Add helper method for checking dkms status.
The dkms status case is special in these tests as the
tests filter to just the status of the module in quesiton.
This means instead of checking a dkms command for the
expected output, it was checking a helper method's output.
This change adds a new helper method specifically for dkms
status and checking its output.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>