fix incomplete deletion of install directory
Since 6fab150cb4daf15d556b7e5860d64b834736f903 the install location
wasn't cleared upon removal due to the test checking for an existing
directory with a relative path. This fixes the issue by checking for
the variable to be empty instead of checking for being a directory.

fixes #276