run_tests: properly track the sig key/hashalgo
Currently we make a few odd assumptions - we track the key field first.
Which may be either empty (kmod build w/o openssl) or missing all
together (aka something has gone horribly wrong). Then based on that we
try and workaround the hashalgo.

That approach was chosen since kmod <26 was buggy (at least the hashalgo
provided was).

Instead - start with a kmod version check (anything half-recent uses
26+) then check the algo, which if unknown tells us to not look at
the empty key. Or if the hash is known and key is missing we can
meaningfully error out.

In the process of this, use modinfo -F instead of the current modinfo |
grep | sed dance.

Side note: the NO_SIGNING_TOOL hunk (and surrounding code) should be
de duplicated into a function or sorts. But that for another day.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>