dkms: skip module signing when hash is unknown
Currently we'll attempt to sign the modules, even when we don't have the
kernel config and/or when CONFIG_MODULE_SIG_HASH is not set.

This serves little purpose, since in either of those cases the kernel
won't check the signature.

Surprise! Turns out Debian 9 does not enable signing, even though it
ships the sign-file tool.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>