CI: remove Debian 9 from the matrix
The distribution has recently been archived and the packages are no
longer reachable... In the default configuration at least - I didn't try
rummaging through the apt config.

AFAICT Debian 9 has gone EOL in 2020, with LTS finishing in 2022. There
is an ELTS still ongoing but that's outside of our concerns... Unless
someone comes up with enticing suggestions that is ;-)

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>