Remove lsb_release
The codebase has moved to use /etc/os-release with a fallback to
lsb_release. That said, the latter has two variants a data file in /etc
or an executable in $PATH.

The in-tree lsb_release is an executable script, although it's installed
outside of $PATH and no distribution (that I can see) adjusts the
environment variable or even uses external lsb_release dependency.

Considering the lack of practical users, let's remove the script. We can
worry about the remaining in-tree lsb-release references later on.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>