make: inline and ban MAN
There is a single instance of the variable, there's no Linux
distributions that I can see overriding the default. Yet most
importantly the variable should really be called MANDIR, to be
consistent with the vast majority of projects out there.

BSD, Solaris and friends tend to use different path, but dkms is not
usually used on those.

As needed we can reinstate the config toggle (as MANDIR really) if
needed.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>