make: reshuffle all/install/tarball target deps
The current all target is a bit bonkers - it cleans all the generated
files, does not explicitly regenerate them and creates a distribution
tarball.

With the latter having the side effect of regenerating the required
files, by having a dependency list.

Similar to the tarball target, install also has an explicit dependency
list of files that need to be generated.

All of this is fairly uncommon to how all the projects (that I've seen)
operate. Namely:

all: all_the_generated_targets

install: all
   install steps

tarball: all

clean:
   foobar ... the target is never a dependency of any of the above

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>