Add 'broken' status
If either a dkms module source, or the symbolic link pointing to
it is missing, the output of `dkms status` will be messed up.
Add a new status called 'broken', which will inform the user
about it in a nicely formatted way.
is_module_added() was modified to not report a module as added,
if not both the source directory and the 'source' symlink exist.
do_autoinstall() and run_match() were modified to handle a broken
status. They skip that particular module/version combo and continue
iterating.
The new function module_is_broken_and_die() was introduced to die
early on a broken module.
Because if in a broken state everything has to be considered volatile,
we always die. User intvervention is required to restore a healthy
environment.
The only exeption is, if only the symbolic link 'source' is missing,
the action 'add' can be used to re-add the module.
The man page was updated with the new 'broken' status.
Tests were added to the test suite.

Signed-off-by: Mart Frauenlob <AllKind@fastest.cc>