Improve 'make' command substitution
'make' may not be necessary the first word in the build command line. If
it is then 'make' substitution fails and '-j' option is not passed to
the 'make' command. For example MAKE directive array may be declared as:

MAKE[0]="source my_environment && make ..."

It causes parallel build to be ignored. The patch imroves parsing of the
build command and substitutes 'make' regardless its position in the
command line.

Change-Id: Id7aefdb5db87821849d5785cdc4b86cb69c259c4
Signed-off-by: Slava Grigorev <alsgnet@gmail.com>