workflows/tests.yml: fix Gentoo testing
Add --autounmask-continue to tell portage to make whatever configuration
changes it deems necessary, and then continue.

Use getbinpkg to fetch prebuilt packages from the binary host to speed
things up a bit. Use binpkg-ignore-signature because getuto is not set
up in the docker container so no trusted keys are set.

Use the generic-uki to avoid generating an initramfs and save a bit of time.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>