Allow overriding /lib/modules via MODDIR
With many distributions moving from /lib/modules to /usr/lib/modules,
allow for the builder to select their option. For backwards
compatibility the default is still /lib/modules.

As result distros for example Arch don't need to patch the project.
Plus this nicely aligns with kmod, which recently allowed for similar
change.

Note: the tests still use the original path. If needed that can be
changed as follow-up.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>