dkms: use saner compression defaults
Currently our compression options vary from those used in upstream
kernel. As of 6.8-rc2 upstream uses (ignoring force/rm input file etc):

 - gzip -9
 - xz --check=crc32 --lzma2=dict=1MiB
   Note: some older kernels lacked crc32 and/or used 2MiB
 - zstd -T0 (aka -3)

With the higher dictionary size (xz) and compression size (20+ zstd)
more memory is required for decompression. Which may be a problem if the
kernel itself is responsible for the decompression (instead of kmod),
since it uses a smaller/limited ram amount.

Reduce the numbers of be compatible with upstream.

Closes: https://github.com/dell/dkms/issues/386

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>